deleted unused pages from Templete/Pages

+++ src/Controller/AssignmentsController.php
@@ -2,6 +2,7 @@
 namespace App\Controller;
 
 use App\Controller\AppController;
+use Cake\ORM\TableRegistry;
 
 /**
  * Assignments Controller
@@ -10,22 +11,6 @@ use App\Controller\AppController;
  */
 class AssignmentsController extends AppController
 {
-
-    /**
-     * Index method
-     *
-     * @return \Cake\Network\Response|null
-     */
-    public function index()
-    {
-        $this->paginate = [
-            'contain' => ['Grades', 'Attachments', 'Users', 'Courses']
-        ];
-        $assignments = $this->paginate($this->Assignments);
-
-        $this->set(compact('assignments'));
-        $this->set('_serialize', ['assignments']);
-    }
 
     /**
      * View method
@@ -37,11 +22,12 @@ class AssignmentsController extends AppController
     public function view($id = null)
     {
         $assignment = $this->Assignments->get($id);
+        $course = $this->Assignments->Courses->get($assignment->course_id);
         $submissions = $this->Assignments->Submissions->find('withAssignments', ['assignment_id' => $id])->contain('Users');
         $members = $this->Assignments->Submissions->Users->find('withCourse', ['course_id' => $assignment->course_id]);
 
-        $this->set(compact('assignment', 'submissions', 'members'));
-        $this->set('_serialize', ['assignment', 'submissions', 'members']);
+        $this->set(compact('assignment', 'submissions', 'members', 'course'));
+        $this->set('_serialize', ['assignment', 'submissions', 'members', 'course']);
     }
 
     /**
@@ -51,19 +37,34 @@ class AssignmentsController extends AppController
      */
     public function add($course_id)
     {
+        //Only the teacher can access here
+        $course = $this->Assignments->Courses->get($course_id);
+        if($course->teacher_id != $this->Auth->user('id')) {
+            $this->Flash->error(__('You are not allowed to access this page.'));
+            return $this->redirect(['action' => 'courseAssignments', $course_id]);
+        }
+
         $assignment = $this->Assignments->newEntity();
-        if ($this->request->is('post')) {
+                if ($this->request->is('post')) {
             $assignment = $this->Assignments->patchEntity($assignment, $this->request->data);
+            
+            $grade = $this->Assignments->Grades->newEntity();
+            $grade->course_id = $course_id;
+            $grade->title = $assignment->title;
+            $grade->description = $assignment->description;
+            if ($this->Assignments->Grades->save($grade)) {
+                $assignment->grade = $grade;
             if ($this->Assignments->save($assignment)) {
+                $this->createAssignmentNotification($assignment);
                 $this->Flash->success(__('The assignment has been saved.'));
 
                 return $this->redirect(['action' => 'courseAssignments', $course_id]);
             }
+            }
             $this->Flash->error(__('The assignment could not be saved. Please, try again.'));
         }
-        $this->set(compact('assignment'));
-        $this->set('_serialize', ['assignment']);
-        $this->set('course_id', $course_id);
+        $this->set(compact('assignment', 'course'));
+        $this->set('_serialize', ['assignment', 'course']);
     }
 
     /**
@@ -75,9 +76,15 @@ class AssignmentsController extends AppController
      */
     public function edit($id = null)
     {
-        $assignment = $this->Assignments->get($id, [
-            'contain' => []
-        ]);
+        $assignment = $this->Assignments->get($id);
+
+        //Only the teacher can access here
+        $course = $this->Assignments->Courses->get($assignment->course_id);
+        if($course->teacher_id != $this->Auth->user('id')) {
+            $this->Flash->error(__('You are not allowed to access this page.'));
+            return $this->redirect(['action' => 'courseAssignments', $course_id]);
+        }
+
         if ($this->request->is(['patch', 'post', 'put'])) {
             $assignment = $this->Assignments->patchEntity($assignment, $this->request->data);
             if ($this->Assignments->save($assignment)) {
@@ -101,22 +108,70 @@ class AssignmentsController extends AppController
     public function delete($id = null)
     {
         $this->request->allowMethod(['post', 'delete']);
+        
         $assignment = $this->Assignments->get($id);
+        //Only the teacher can access here
+        $course = $this->Assignments->Courses->get($assignment->course_id);
+        if($course->teacher_id != $this->Auth->user('id')) {
+            $this->Flash->error(__('You are not allowed to access this page.'));
+            return $this->redirect(['action' => 'courseAssignments', $course_id]);
+        }
+         
         if ($this->Assignments->delete($assignment)) {
             $this->Flash->success(__('The assignment has been deleted.'));
         } else {
             $this->Flash->error(__('The assignment could not be deleted. Please, try again.'));
         }
 
-        return $this->redirect(['action' => 'index']);
+        return $this->redirect(['action' => 'courseAssignments', $course->id]);
     }
 
     public function courseAssignments($course_id)
     {
-        $assignments = $this->paginate($this->Assignments->find('withCourse', ['course_id' => $course_id]));
+        //Only admin and course members can access.
+        if ($this->Auth->user('role') != 1 && !$this->isMember($course_id, $this->Auth->user('id'))) {
+             $this->Flash->error(__('You are not allowed to access this page.'));
+            return $this->redirect(['controller' => 'Courses', 'action' => 'view', $course_id]);
+        }
 
-        $this->set(compact('assignments'));
-        $this->set('_serialize', ['assignments']);
-        $this->set('course_id', $course_id);       
+        $assignments = $this->paginate($this->Assignments->find('withCourse', ['course_id' => $course_id]));
+        $course = $this->Assignments->Courses->get($course_id);
+        
+        $this->set(compact('assignments', 'course'));
+        $this->set('_serialize', ['assignments', 'course']);
+    }
+
+    private function createAssignmentNotification($assignment)
+    {
+        $Notifications = TableRegistry::get('Notifications');
+        // Create notification and messages to all members.
+        $notification = $Notifications->newEntity();
+        $notification->type = 3;// Added assignment
+        $notification->variable1 = $assignment->id;
+        $notification->date = date("Y-m-d H:i:s");
+        $notification->sender_id = $this->Auth->user('id');
+                
+        if ($Notifications->save($notification)) {
+            $Messages = TableRegistry::get('Messages');
+            $members = $this->Assignments->Courses->Users->find('withCourse', ['course_id' => $assignment->course_id]);
+            foreach($members as $member) {
+                $message = $Messages->newEntity();
+                $message->notification_id = $notification->id;
+                $message->receiver_id = $member->id;
+                $message->isRead = 0;
+                $Messages->save($message);
+            }
+        }
+    }
+
+    private function isMember($course_id, $user_id)
+    {
+        $query = $this->Assignments->Courses->CoursesUsers->find('all')
+            ->where(['course_id' => $course_id, 'user_id' => $user_id]);
+        if ($query->count() >= 1) {
+            return true;
+        } else {
+            return false;
+        }
     }
 }
